Remove all deprecations and majority of private API usage
In Ember 3.6 we introduced a public way of getting meta data about a
route via a `RoutInfo`. A `RouteInfo` is a public version of handlerInfo.
We also introduced new router events that see things such as aborts, retries,
redirects etc. With these 2 changes we are able to remove the vast majority of
the private API usage in this addon. The only place we do need the private api
usage is looking up the actual route instance. We cannot use `getOwner` because
the owner is not going to have visibility into engines. Also, it appears we need
to schedule the `prefetch` calls into the `actions` queue to get the same timing
semantics that we had before, which was chaining off of some internal promise in
the router.